+++ src/models/game.h
@@ -37,8 +37,8 @@ class Game {
 		void getScalar();
       		std::vector<Frame*> getFrames();
         	int getHome();
-		int getX();
-		int getY();
+		double getX();
+		double getY();
 
 };
 #endif
